Add a benchmark for batch_opening
* Move all fixutres in conftest.py to fixtures.py so that they can be
reused between `benchmark` and `tests`
* Add a benchmark for `batch_opening`
* Add a README on how to add/run benchmarks
