+++ benchmark/conftest.py
@@ -1,0 +1,1 @@
+from tests.fixtures import *  # noqa: F403 F401
